requirements.txt
Flask==2.2.2
python-dotenv==0.21.1
watchdog==2.2.1
Flask-SQLAlchemy==3.0.3
mysqlclient==2.1.1
Flask-Migrate==4.0.4
APScheduler==3.10.0
python-dateutil==2.8.2
Flask-Mail==0.9.1